SUIKA-Net/requirements_full.txt
timm>=1.0.22
torch==2.5.1+cu121
torchvision==0.20.1+cu121
pandas>=2.3.3
ultralytics>=8.4.37
rembg>=2.0.69
onnxruntime-gpu>=1.23.2
huggingface_hub>=1.10.1
opencv-python>=4.13.0
opencv-python-headless>=4.13.0
numpy>=2.2.6
Pillow>=12.2.0
PyYAML>=6.0.3
